+++ server/ts/util.ts
@@ -2,9 +2,8 @@ import { ImageMagick } from '@imagemagick/magick-wasm';
 import { MagickFormat } from '@imagemagick/magick-wasm/magick-format';
 import * as fs from 'fs-extra';
 import * as path from 'path';
-import { Config } from './config';
 import { DirectoryStructure } from './globals';
-import { DataDefinition, GameDefinition } from '../../shared/types';
+import { DataDefinitionBase } from '../../shared/types';
 import { Mission } from './mission';
 
 export class Util {
@@ -298,11 +297,11 @@ export class Util {
 	}
 
 	/** Choose data playable by the given datablockCompatibility i.e filters the data whose datablocks are compatible with that of datablockCompatibility */
-	static chooseDataByDatablockCompatibility<T extends string>(games: DataDefinition<T>[], datablockCompatibility: Mission["datablockCompatibility"]) {
+	static chooseDataByDatablockCompatibility<T extends DataDefinitionBase>(games: T[], datablockCompatibility: Mission["datablockCompatibility"]) {
 		// PQ levels can only run on PQ
 		// MBW levels can run on both PQ and MBW
 		// MBG levels can run on all
-		let selected: DataDefinition<T>[] = [];
+		let selected: T[] = [];
 		let compatibilityLevel = { mbg: 0, mbw: 1, pq: 2 }[datablockCompatibility];
 		for (const game of games) {
 			let gameCompatibilityLevel = { mbg: 0, mbw: 1, pq: 2 }[game.datablockCompatibility];